Create SimAtClientPoi.js
In conclusion, it simulates the flow of customers in a bank over a period of 8 hours. In that sense, customers arrive according to a Poisson distribution. In which they are served by tellers at available tellers and the status of the tellers and queue is shown every minute. With that, we conclude which cashier was the busiest in terms of customers served and which cashiers were busiest in terms of service time.